feat(iam): add role association commands for API key and user roles, reorganize file structure
